Split history into a separate function
