+++ src/models/api/HbbResponse.ts
@@ -1,4 +1,3 @@
-import { TimestampValueResponse } from './TimestampValueResponse';
 
 export type HbbResponse = {
   // Amount of HBB staked (total staked in the staking pool)
@@ -9,10 +8,6 @@ export type HbbResponse = {
   numberOfStakers: number;
   // Total HBB issued (current supply from HBB mint account)
   issued: number;
-  // HBB price through history
-  priceHistory: TimestampValueResponse[];
   // Number of current HBB holders
   numberOfHolders: number;
-  // Number of HBB holders through history
-  holdersHistory: TimestampValueResponse[];
 };
