+++ src/models/api/LoansResponse.ts
@@ -1,11 +1,8 @@
-import { TimestampValueResponse } from './TimestampValueResponse';
 import { PercentileResponse } from './PercentileResponse';
 
 export type LoansResponse = {
   // Number of current loans
   total: number;
-  // Number of loans through history
-  history: TimestampValueResponse[];
   // Loan amount distribution
   distribution: PercentileResponse[];
   // Largest loan in the borrowing market
