+++ src/models/api/UsdhResponse.ts
@@ -1,4 +1,3 @@
-import { TimestampValueResponse } from './TimestampValueResponse';
 import { PriceResponse } from './PriceResponse';
 import { PercentileResponse } from './PercentileResponse';
 
@@ -7,8 +6,6 @@ export type UsdhResponse = {
   issued: number;
   // Total in stability pool / stability provided (stablecoin deposited in stability pool)
   stabilityPool: number;
-  // Total USDH issued through history
-  history: TimestampValueResponse[];
   // Distribution of USDH in the stability pool
   stabilityPoolDistribution: PercentileResponse[];
   // Price and liquidity pool values obtained from Saber
